# Improve performance of Java agent