--- a/scavenger-agent-java/src/integrationTest/java/integrationTest/javaagent/BenchmarkTest.java
+++ b/scavenger-agent-java/src/integrationTest/java/integrationTest/javaagent/BenchmarkTest.java
@@ -2,13 +2,11 @@ package integrationTest.javaagent;
 
 import integrationTest.support.AgentBenchmarkExtension;
 import integrationTest.support.AgentRunner;
-import lombok.extern.java.Log;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.TestTemplate;
 import org.junit.jupiter.api.extension.ExtendWith;
 
-@Log
 @Disabled("Comment/remove this annotation to run benchmarks")
 @ExtendWith(AgentBenchmarkExtension.class)
 public class BenchmarkTest {