--- a/scavenger-agent-java/src/integrationTest/java/integrationTest/javaagent/InvocationTest.java
+++ b/scavenger-agent-java/src/integrationTest/java/integrationTest/javaagent/InvocationTest.java
@@ -47,28 +47,8 @@ import com.navercorp.scavenger.model.InitConfigResponse;
 import com.navercorp.scavenger.model.PublicationResponse;
 
 @ExtendWith(AgentIntegrationTestContextProvider.class)
-@ExtendWith(GrpcMockExtension.class)
 @DisplayName("invocation track test")
-public class InvocationTest {
-    private static WireMockServer wireMockServer;
-    private static ManagedChannel channel;
-
-    @BeforeAll
-    static void setUp() {
-        wireMockServer = new WireMockServer(new WireMockConfiguration().dynamicPort());
-        wireMockServer.start();
-        WireMock.configureFor(wireMockServer.port());
-
-        channel = ManagedChannelBuilder.forAddress("localhost", GrpcMock.getGlobalPort())
-            .usePlaintext()
-            .build();
-    }
-
-    @AfterAll
-    static void tearDown() {
-        Optional.ofNullable(wireMockServer).ifPresent(WireMockServer::shutdown);
-        Optional.ofNullable(channel).ifPresent(ManagedChannel::shutdownNow);
-    }
+public class InvocationTest extends AbstractWireMockTest {
 
     @TestTemplate
     @DisplayName("it tracks correctly")