--- a/scavenger-agent-java/src/integrationTest/java/integrationTest/javaagent/InvocationTest.java
+++ b/scavenger-agent-java/src/integrationTest/java/integrationTest/javaagent/InvocationTest.java
@@ -68,10 +68,8 @@ public class InvocationTest extends AbstractWireMockTest {
     @DisplayName("it sends publication correctly")
     void send(AgentRunner agentRunner) throws Exception {
         // given
-        Properties properties = new Properties();
-        properties.setProperty("serverUrl", "http://localhost:" + wireMockServer.port());
-        properties.setProperty("schedulerInitialDelayMillis", "0");
-        agentRunner.setConfig(properties);
+        agentRunner.setConfigProperty("serverUrl", "http://localhost:" + wireMockServer.port());
+        agentRunner.setConfigProperty("schedulerInitialDelayMillis", "0");
 
         givenThat(
             get(V5_INIT_CONFIG + "?licenseKey=")