--- a/scavenger-agent-java/src/integrationTest/java/integrationTest/javaagent/ScanTest.java
+++ b/scavenger-agent-java/src/integrationTest/java/integrationTest/javaagent/ScanTest.java
@@ -5,6 +5,7 @@ import static com.github.tomakehurst.wiremock.client.WireMock.givenThat;
 import static com.github.tomakehurst.wiremock.client.WireMock.okJson;
 import static com.navercorp.scavenger.model.Endpoints.Agent.V5_INIT_CONFIG;
 import static integrationTest.util.AgentLogAssertionUtil.assertSampleAppOutput;
+import static integrationTest.util.AgentLogAssertionUtil.extractFromMatchingLogLines;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.grpcmock.GrpcMock.calledMethod;
 import static org.grpcmock.GrpcMock.getGlobalPort;
@@ -13,41 +14,29 @@ import static org.grpcmock.GrpcMock.unaryMethod;
 import static org.grpcmock.GrpcMock.verifyThat;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import java.lang.reflect.Method;
-import java.util.Optional;
-import java.util.Properties;
+import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import org.grpcmock.GrpcMock;
-import org.grpcmock.junit5.GrpcMockExtension;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
+import com.navercorp.scavenger.javaagent.collecting.CodeBaseScanner;
+import com.navercorp.scavenger.javaagent.collecting.InvocationTracker;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.TestTemplate;
 import org.junit.jupiter.api.extension.ExtendWith;
 
-import com.github.tomakehurst.wiremock.WireMockServer;
-import com.github.tomakehurst.wiremock.client.WireMock;
-import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
 import com.google.protobuf.util.JsonFormat;
 import integrationTest.support.AgentIntegrationTestContextProvider;
 import integrationTest.support.AgentRunner;
-import integrationTest.util.AgentLogAssertionUtil;
-import io.grpc.ManagedChannel;
-import io.grpc.ManagedChannelBuilder;
 import sample.app.NotServiceClass;
+import sample.app.SampleApp;
+import sample.app.SampleAspect;
 import sample.app.SampleService1;
-import sample.app.excluded.NotTrackedClass;
-import sample.app.excluded.NotTrackedClass2;
+import sample.app.SampleService2;
 
 import com.navercorp.scavenger.model.GetConfigResponse;
 import com.navercorp.scavenger.model.GrpcAgentServiceGrpc;
 import com.navercorp.scavenger.model.InitConfigResponse;
 import com.navercorp.scavenger.model.PublicationResponse;
-
-import sample.app.regex.ExcludeRegexClass1;
-import sample.app.regex.ExcludeRegexClass2;
 
 @ExtendWith(AgentIntegrationTestContextProvider.class)
 @DisplayName("codebase scan test")
@@ -61,12 +50,39 @@ public class ScanTest extends AbstractWireMockTest {
 
         // then
         assertSampleAppOutput(stdout);
-        assertThat(stdout).matches(scanned(SampleService1.class.getMethod("doSomething", int.class)));
-        assertThat(stdout).matches(scanned(NotServiceClass.class.getMethod("doSomething", int.class)));
-        assertThat(stdout).doesNotMatch(scanned(NotTrackedClass.class.getMethod("doSomething")));
-        assertThat(stdout).doesNotMatch(scanned(NotTrackedClass2.class.getMethod("doSomething")));
-        assertThat(stdout).doesNotMatch(scanned(ExcludeRegexClass1.class.getMethod("doSomething")));
-        assertThat(stdout).doesNotMatch(scanned(ExcludeRegexClass2.class.getMethod("doSomething")));
+        List<String> scannedMethods = extractFromMatchingLogLines(stdout, CodeBaseScanner.class, "[scavenger] ", " is scanned");
+        assertThat(scannedMethods).containsExactlyInAnyOrder(
+                "sample.app.NotServiceClass()",
+                "sample.app.NotServiceClass.doNothing()",
+                "sample.app.NotServiceClass.doSomething(int)",
+                "sample.app.SampleApp(sample.app.SampleService1)",
+                "sample.app.SampleApp.add(int,int)",
+                "sample.app.SampleApp.main(java.lang.String[])",
+                "sample.app.SampleApp.postConstruct()",
+                "sample.app.SampleAspect()",
+                "sample.app.SampleAspect.aroundSampleService(org.aspectj.lang.ProceedingJoinPoint)",
+                "sample.app.SampleAspect.logAspectLoaded()",
+                "sample.app.SampleService1(sample.app.SampleService2)",
+                "sample.app.SampleService1.doSomething(int)",
+                "sample.app.SampleService2()",
+                "sample.app.SampleService2.doSomething(int)"
+        );
+    }
+
+    @TestTemplate
+    @DisplayName("it installs advice correctly")
+    void advice(AgentRunner agentRunner) throws Exception {
+        String stdout = agentRunner.call();
+
+        List<String> installedAdvice = extractFromMatchingLogLines(stdout, InvocationTracker.class,
+                "[scavenger] Advice on ", " is installed");
+
+        assertThat(installedAdvice).containsExactlyInAnyOrder(
+                SampleApp.class.getName(),
+                SampleService1.class.getName(),
+                SampleAspect.class.getName(),
+                SampleService2.class.getName(),
+                NotServiceClass.class.getName());
     }
 
     @TestTemplate
@@ -109,13 +125,6 @@ public class ScanTest extends AbstractWireMockTest {
             calledMethod(GrpcAgentServiceGrpc.getSendCodeBasePublicationMethod())
                 .withStatusOk()
                 .withRequest(pub -> pub.getEntryCount() == getMethodsCount(stdout)));
-    }
-
-    private static Pattern scanned(Method method) {
-        String[] split = method.toString().split(" ");
-        String signature = split[split.length - 1];
-        return AgentLogAssertionUtil.logPattern("com.navercorp.scavenger.javaagent.collecting.CodeBaseScanner",
-            "[scavenger] " + signature + " is scanned");
     }
 
     private static int getMethodsCount(String stdout) {