--- a/scavenger-agent-java/src/integrationTest/java/integrationTest/support/AgentBenchmarkExtension.java
+++ b/scavenger-agent-java/src/integrationTest/java/integrationTest/support/AgentBenchmarkExtension.java
@@ -1,0 +1,14 @@
+package integrationTest.support;
+
+public class AgentBenchmarkExtension extends AgentIntegrationTestContextProvider {
+
+    @Override
+    protected Class<?> getTestAppMainClass() {
+        return org.openjdk.jmh.Main.class;
+    }
+
+    @Override
+    protected String getScavengerConfigPath() {
+        return "scavenger-jmh.conf";
+    }
+}