--- a/scavenger-agent-java/src/integrationTest/java/integrationTest/support/AgentRunner.java
+++ b/scavenger-agent-java/src/integrationTest/java/integrationTest/support/AgentRunner.java
@@ -1,37 +1,40 @@
 package integrationTest.support;
 
+import static java.util.stream.Collectors.joining;
+
+import lombok.Setter;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
-import java.util.Properties;
+import java.util.Map;
 import java.util.concurrent.Callable;
-import java.util.stream.Collectors;
 
 public class AgentRunner implements Callable<String> {
     private final String javaPath;
     private final String classpath;
+    private final String mainClass;
     private final String agentPath;
-
+    private final Map<String, String> configProps = new HashMap<>();
+    @Setter
+    private boolean shouldLogOutput;
+    @Setter
     private String configFilePath;
-    private String cmdLineArgs;
 
-    public AgentRunner(String javaPath, String classpath, String agentPath) {
+    public AgentRunner(String javaPath, String classpath, Class<?> mainClass, String agentPath, String configFilePath) {
         this.javaPath = javaPath;
         this.classpath = classpath;
+        this.mainClass = mainClass.getName();
         this.agentPath = agentPath;
-    }
-
-    public void setConfigFilePath(String configFilePath) {
         this.configFilePath = configFilePath;
     }
 
-    public void setConfig(Properties properties) {
-        cmdLineArgs = properties.entrySet().stream()
-            .map(entry -> entry.getKey() + "=" + entry.getValue())
-            .collect(Collectors.joining(";"));
+    public void setConfigProperty(String key, String value) {
+        configProps.put(key, value);
     }
 
     @Override
@@ -47,33 +50,44 @@ public class AgentRunner implements Callable<String> {
     }
 
     private List<String> buildCommand() {
-        String cp = classpath.endsWith(":") ? classpath.substring(0, classpath.length() - 2) : classpath;
-
         List<String> command = new ArrayList<>();
         command.add(javaPath);
-        if (cmdLineArgs != null) {
-            command.add("-javaagent:" + agentPath + "=" + cmdLineArgs);
-        } else {
-            command.add("-javaagent:" + agentPath);
+        String agentCmd = "-javaagent:" + agentPath;
+        if (!configProps.isEmpty()) {
+            agentCmd += "=" + cmdLineArgsToString();
         }
+        command.add(agentCmd);
         command.add("-cp");
+        String cp = classpath.endsWith(":") ? classpath.substring(0, classpath.length() - 2) : classpath;
         command.add(cp);
         command.add("-Djava.util.logging.config.file=src/integrationTest/resources/logging.properties");
         command.add("-Duser.language=en");
         command.add("-Duser.country=US");
-        if (configFilePath == null) {
-            command.add("-Dscavenger.configuration=src/integrationTest/resources/scavenger.conf");
-        } else if (!configFilePath.isEmpty()) {
-            command.add("-Dscavenger.configuration=" + configFilePath);
+        if (configFilePath != null) {
+            command.add("-Dscavenger.configuration=src/integrationTest/resources/" + configFilePath);
         }
-        command.add("sample.app.SampleApp");
+        command.add(mainClass);
         System.out.println("Launching SampleApp with the command: " + command);
         return command;
     }
 
-    private static String collectProcessOutput(InputStream inputStream) throws IOException {
+    private String cmdLineArgsToString() {
+        return configProps.entrySet().stream()
+                .map(entry -> entry.getKey() + "=" + entry.getValue())
+                .collect(joining(";"));
+    }
+
+    private String collectProcessOutput(InputStream inputStream) throws IOException {
+        StringBuilder output = new StringBuilder();
         try (BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream))) {
-            return reader.lines().collect(Collectors.joining("\n"));
+            String line;
+            while ((line = reader.readLine()) != null) {
+                if (shouldLogOutput) {
+                    System.out.println(line);
+                }
+                output.append(line).append("\n");
+            }
         }
+        return output.toString();
     }
 }