--- a/scavenger-agent-java/src/integrationTest/java/integrationTest/util/AgentLogAssertionUtil.java
+++ b/scavenger-agent-java/src/integrationTest/java/integrationTest/util/AgentLogAssertionUtil.java
@@ -2,6 +2,9 @@ package integrationTest.util;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class AgentLogAssertionUtil {
@@ -21,5 +24,24 @@ public class AgentLogAssertionUtil {
 
     public static Pattern logPattern(String location, String text) {
         return Pattern.compile("[\\s\\S]*(INFO).*" + Pattern.quote(location) + ".*" + Pattern.quote(text) + "[\\s\\S]*");
+    }
+
+    public static List<String> extractFromMatchingLogLines(String multiLineLogOutput, Class<?> loggingClass, String matchingPrefix, String matchingSuffix) {
+        return extractFromMatchingLogLines(multiLineLogOutput, loggingClass, Pattern.quote(matchingPrefix) + "(.*)" + Pattern.quote(matchingSuffix));
+    }
+
+    public static List<String> extractFromMatchingLogLines(String multiLineLogOutput, Class<?> loggingClass, String messageRegexWithCaptureGroup) {
+        List<String> result = new ArrayList<>();
+        String loggerName = loggingClass.getName();
+        String logLineRegex = ".*?\\[.*?] " + Pattern.quote(loggerName) + " - " + messageRegexWithCaptureGroup;
+        Pattern logLinePattern = Pattern.compile(logLineRegex, Pattern.MULTILINE);
+        Matcher matcher = logLinePattern.matcher(multiLineLogOutput);
+        while (matcher.find()) {
+            String capturedValue = matcher.group(1);
+            if (capturedValue != null && !capturedValue.isEmpty()) {
+                result.add(capturedValue);
+            }
+        }
+        return result;
     }
 }