--- a/scavenger-agent-java/src/integrationTest/java/sample/app/NotServiceClass.java
+++ b/scavenger-agent-java/src/integrationTest/java/sample/app/NotServiceClass.java
@@ -8,4 +8,8 @@ public class NotServiceClass {
     public static void doSomething(int p1) {
         log.info("Doing something " + p1);
     }
+
+    public static void doNothing() {
+        // Intentionally left blank
+    }
 }