--- a/scavenger-agent-java/src/integrationTest/java/sample/app/excluded/jmh/BenchmarkRunner.java
+++ b/scavenger-agent-java/src/integrationTest/java/sample/app/excluded/jmh/BenchmarkRunner.java
@@ -1,0 +1,27 @@
+package sample.app.excluded.jmh;
+
+import org.openjdk.jmh.annotations.Benchmark;
+import org.openjdk.jmh.annotations.BenchmarkMode;
+import org.openjdk.jmh.annotations.Fork;
+import org.openjdk.jmh.annotations.Measurement;
+import org.openjdk.jmh.annotations.Mode;
+import org.openjdk.jmh.annotations.OutputTimeUnit;
+import org.openjdk.jmh.annotations.Threads;
+import org.openjdk.jmh.annotations.Warmup;
+import sample.app.NotServiceClass;
+
+import java.util.concurrent.TimeUnit;
+
+@BenchmarkMode(Mode.AverageTime)
+@OutputTimeUnit(TimeUnit.NANOSECONDS)
+@Fork(2)
+@Threads(4)
+@Warmup(iterations = 2)
+@Measurement(iterations = 3)
+public class BenchmarkRunner {
+
+    @Benchmark
+    public void bench() {
+        NotServiceClass.doNothing();
+    }
+}