--- a/scavenger-agent-java/src/integrationTest/resources/scavenger-jmh.conf
+++ b/scavenger-agent-java/src/integrationTest/resources/scavenger-jmh.conf
@@ -1,0 +1,7 @@
+appName = SampleApp
+packages = sample
+methodVisibility = protected
+excludePackages = sample.app.excluded
+codeBase = build/classes/java/integrationTest
+debugMode = false
+schedulerInitialDelayMillis = 1000000